+++ application/controllers/requerimentos.php
@@ -21,7 +21,14 @@ class Requerimentos extends MY_Controller
     {
         $this->data['requerimentos'] = $this->requerimento_model->get_requerimentos_with_bairros();
 
-        $this->load_view('requerimentos/listar_requerimentos');
+        $this->load_view('requerimentos/listar_requerimentos', TRUE);
+    }
+
+    public function meus_requerimentos()
+    {
+        $this->data['requerimentos'] = $this->requerimento_model->get_meus_requerimentos_with_bairros($_SESSION['id_user']);
+
+        $this->load_view('requerimentos/meus_requerimentos');
     }
 
     public function cadastrar_requerimento()
@@ -32,7 +39,7 @@ class Requerimentos extends MY_Controller
         $config['max_width'] = '2048';
         $config['max_height'] = '1024';
         $config['encrypt_name'] = TRUE;
-        
+
         $this->load->library('upload', $config);
 
         $this->form_validation->set_rules($this->requerimento_model->validation);
@@ -40,9 +47,9 @@ class Requerimentos extends MY_Controller
         $this->data['bairros'] = $this->bairros_model->get_all();
         $this->data['requerentes'] = $this->requerente_model->get_vereadores();
         $this->data['cats_requerimento'] = $this->categorias_requerimento_model->get_all();
-        
+
         if ($this->form_validation->run()==TRUE):
-            $data = elements(array('descricao','id_bairro','id_rua','cat_requerimento','id_requerente'),$this->input->post());
+            $data = elements(array('descricao','id_bairro','id_rua','cat_requerimento','id_requerente','id_criador'),$this->input->post());
             $data['data_requerimento'] = date('Y-m-d');
 
             $i = 1;
@@ -69,7 +76,7 @@ class Requerimentos extends MY_Controller
             {
                 $this->requerimento_model->insert($data);
                 generate_charts();
-                
+
                 $this->session->set_userdata('requerimento_cadastrado','Requerimento cadastrado com sucesso!');
 
                 redirect('requerimentos/cadastrar_requerimento');
@@ -79,17 +86,15 @@ class Requerimentos extends MY_Controller
         $this->load_view('requerimentos/cadastrar_requerimento');
     }
 
-    public function editar_requerimento()
-    {      
-        $id = $this->uri->segment(3);
-        
+    public function editar_requerimento($id)
+    {
         $config['upload_path'] = './uploads/';
         $config['allowed_types'] = 'gif|jpg|jpeg|png';
         $config['max_size'] = '2048';
         $config['max_width'] = '2048';
         $config['max_height'] = '1024';
         $config['encrypt_name'] = TRUE;
-        
+
         $this->load->library('upload', $config);
 
         $this->form_validation->set_rules($this->requerimento_model->validation);
@@ -126,29 +131,30 @@ class Requerimentos extends MY_Controller
             {
                 $this->requerimento_model->update($this->input->post('id'),$data);
                 generate_charts();
-                
+
                 $this->session->set_userdata('requerimento_editado','Requerimento editado com sucesso!');
 
                 redirect('requerimentos/editar_requerimento/'.$this->input->post('id'));
             }
         endif;
-        
+
         $this->data['requerimento'] = $this->requerimento_model->get($id);
-                
+
         if ($this->data['requerimento']->id_rua != 0)
             $this->data['ruas'] = $this->cidades_model->getRuas($this->data['requerimento']->id_bairro);
 
         $this->load_view('requerimentos/editar_requerimento');
     }
 
-    public function excluir_requerimento()
+    public function excluir_requerimento($id)
     {
-        $id = $this->uri->segment(3);
+        if ($_SESSION['autorizacao'==AUTORIZACAO_ADMINISTRADOR])
+        {
+            $this->requerimento_model->delete($id);
+            generate_charts();
 
-        $this->requerimento_model->delete($id);
-        generate_charts();
-        
-        $this->session->set_userdata('requerimento_excluido','Requerimento excluído com sucesso!');
-        redirect('requerimentos/listar_requerimentos');
+            $this->session->set_userdata('requerimento_excluido','Requerimento excluído com sucesso!');
+            redirect('requerimentos/listar_requerimentos');
+        }
     }
 }
